\name{shinyCombine}
\docType{methods}
\alias{shinyCombine}
\title{Combining two shinyMethylSet objects}
\description{
Combining two shinyMethylSet objects
}

\usage{
shinyCombine(shinyMethylSet1, shinyMethylSet2)
}

\arguments{
  \item{shinyMethylSet1}{a \code{shinyMethylSet}} 
  \item{shinyMethylSet2}{a \code{shinyMethylSet}} 
}


\details{
  Either a \code{RGChannelSet} is transformed into a \code{shinyMethylSet}, or a \code{GenomicRatioSet} is transformed to a \code{shinyMethylSet}
}

\author{Jean-Philippe Fortin \email{jfortin@jhsph.edu}}

\seealso{
  \code{\link{shinyMethylSet}} for the input and output objects
}

\keyword{methods}

