\name{shinyMethyl-internal}
\alias{extractXYMedians}
\alias{returnPCScores}
\alias{mergeExtractedData}
\alias{mergeListOfVectors}
\alias{mergeListOfMatrices}
\alias{returnSampleDistance}
\alias{returnSummary450k}
\alias{returnSummaryRGSet450k}
\title{Internal shinyMethyl Functions}
\description{
  Internal shinyMethyl functions
}
\details{
  These are not to be called by the user.
}
\keyword{ internal }

